refactor: consolidate VideoItem and YouTubeVideoMetadata into Controller types
