+++ src/pages/Controller/components/Mixer/types/index.ts
@@ -1,9 +1,9 @@
-import type { YouTubeVideoMetadata } from "@/types";
+import type { VideoItem } from "@/pages/Controller/types/videoItem";
 
 export interface MixerAPI {
   setCrossfader: (value: number) => void;
-  setPreparedVideo: (video: YouTubeVideoMetadata | string) => void;
-  getPreparedVideo: () => YouTubeVideoMetadata | null;
+  setPreparedVideo: (video: VideoItem | string) => void;
+  getPreparedVideo: () => VideoItem | null;
   setIsAudioOn: (deckId: number, state: boolean) => void;
   setOpacityValue: (deckId: number, value: number) => void;
 }
