+++ src/pages/Controller/types/videoItem.ts
@@ -1,0 +1,5 @@
+export interface VideoItem {
+  id: string;
+  title?: string;
+  start?: number;
+}
